Разделила на файлы вход в программу: python main.py annotation.csv
